cleanup go project
