+++ go/go.mod
@@ -1,5 +1,3 @@
-module "go"
+module example
 
 go 1.21
-
-require golang.org/x/exp v0.0.0-20230522175609-2e198f4a06a1
