+++ go/interfaces.go
@@ -1,0 +1,5 @@
+package main
+
+type FuelDependent interface {
+	refuel(liter float32)
+}
